// frontend/src/types/user.ts
export interface User {
  id: number;
  username: string;
  email: string;
  role: string;
  isTemporaryPassword: boolean;
  country: string;
  countryCode: string;
  countryId: number;
  department: string;
  departmentId: number;
  designation: string;
  name: string;
  address: string;
  phone: string;
  managerName: string;
  managerDesignation: string;
}